Add file to a module
